zephyr: Introduce WITH_ZEPHYR_LIB option
Introduce an option for when we build libmetal as part of Zephyr itself
instead of as a standalone library.  This is for when libmetal is part
of the Zephyr code base and being built as part of a standard Zephyr
application.  The standalone (WITH_ZEPHYR_LIB=OFF) case we first build
zephyr and use that version to build a libmetal.a and associated
headers.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>