log: temporary patch to highlight weak issue on zephyr
This patch is for zephyr .
Aim of this patch is to declare a weak and strong metal_default_log_handler
function.
In libmetal.a we have (readelf -s ./lib/libmetal.a | grep log)
    52: 00000000     0 NOTYPE  GLOBAL DEFAULT  UND metal_default_log_handler
File: ./lib/libmetal.a(log.c.obj)
     1: 00000000     0 FILE    LOCAL  DEFAULT  ABS log.c
    26: 00000001    96 FUNC    WEAK   DEFAULT    4 metal_default_log_handler
    47: 00000000     0 NOTYPE  GLOBAL DEFAULT  UND metal_default_log_handler
File: ./lib/libmetal.a(log.c.obj)
     1: 00000000     0 FILE    LOCAL  DEFAULT  ABS log.c
    26: 00000001    80 FUNC    GLOBAL DEFAULT    4 metal_default_log_handler

Is is normal that both are include in lib?
as weak function is linked first it will be used by application.


Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@st.com>